Add another crate, add a generalized build script

+++ wasm/fluid-experimental-wasm-add/src/lib.rs
@@ -1,7 +1,7 @@
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: f64, right: f64) -> f64{
     left + right
 }
 
@@ -10,6 +10,6 @@ pub mod test {
     use crate::add;
     #[test]
     pub fn adds_correctly() {
-        assert_eq!(add(1, 2), 3);
+        assert_eq!(add(1 as f64, 2 as f64), 3 as f64);
     }
 }
